[TM-3359] move files constants
